pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-local"
version = "0.1.0"
description = "Offline face grouping and person labeling with Coral Edge TPU support"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "numpy>=1.24,<2.0",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "PySide6>=6.6",
    "SQLAlchemy>=2.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
tflite = [
    "ai-edge-litert>=1.0",
    "pycoral>=2.0",
]
dev = [
    "pytest>=7.4",
    "pytest-qt>=4.2",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
face-local = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
